screen테스트 디렉토리 변경
COIN-25
