+++ test/screen/RankScreenTest.java
@@ -1,4 +1,4 @@
-package test.screen;
+package screen;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
